requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.26
python-dotenv==0.19.1
pandas==1.5.3
mysqlclient==2.2.0
pymysql==1.0.2
Flask-MySQLdb==0.2.0
loguru==0.5.3
